Refactered code
