+++ portfolio/src/views/AboutMe.jsx
@@ -1,0 +1,17 @@
+const AboutMe = () => {
+    return (
+      <div>
+        <p>
+          Hello, it is a pleasure to meet you! I'm a Junior Full Stack Developer
+          with a strong foundation in the MERN stack, specializing in backend
+          development. I have experience building scalable APIs, managing
+          databases, and handling authentication. While my strengths lie in
+          backend development, I also have a solid grasp of frontend technologies
+          and am continuously improving to become more proficient. I’m eager to
+          expand my skill set, take on new challenges, and contribute to impactful
+          projects.
+        </p>
+      </div>
+    );
+  };
+  export default AboutMe;
