+++ portfolio/src/views/HomePage.jsx
@@ -1,0 +1,8 @@
+const HomePage = () => {
+  return (
+    <div>
+      <h1>Home Page</h1>
+    </div>
+  );
+};
+export default HomePage;
