+++ portfolio/src/views/ProjectsPage.jsx
@@ -1,0 +1,8 @@
+const ProjectsPage = () => {
+  return (
+    <div>
+      <h1>Projects Page</h1>
+    </div>
+  );
+};
+export default ProjectsPage;
